+++ internal/cli/atlas/cloudprovider/accessroles/aws/create.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package accessroles
+package aws
 
 import (
 	"github.com/mongodb/mongocli/internal/cli"
