+++ internal/cli/atlas/cloudprovider/accessroles/aws/create_test.go
@@ -14,7 +14,7 @@
 
 // +build unit
 
-package accessroles
+package aws
 
 import (
 	"testing"
