+++ internal/cli/atlas/cloudprovider/accessroles/description.go
@@ -11,8 +11,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package aws
+package accessroles
 
 const (
-	aws = "Manage AWS IAM role access in Atlas."
+	accessRole = "Manage AWS IAM role access in Atlas."
 )
